Split storage2 into files
