+++ src/storage2/mod.rs
@@ -1,0 +1,4 @@
+mod json;
+mod storage;
+
+pub use storage::{MessageId, Storage};
